Bring dual FileMgr instances online close #549 close #550
book.py:
incorporate global signal instance into BookC.

book_ease.glade:
remove the file manager widgets.
remove the playlist opener/creator widgets.

book_ease.py:
remove all bookmark related code and move it to its own module.
MainWindow account for accessing the file manager visibility control via
file_manager_pane.
main() has to create the global signals.
main() has altered the initialization sequence to account for the gtk
templates.

book_reader.py:
BookReader now listens for openbook signals from the global transmitter
BookReader no longer manages the NewBookOpener and ExistingBookOpener
classes.
the NewBookOpener and ExistingBookOpener classes have been removed from
the module and moved to file_manager_view.py
convert more string paths to use pathlib.Path.

file_mgr.py:
account for BookMark being moved to its own module.
add media file awareness and testing to FileList and its nested iterator
class.
clean up directory navigation methods in FileMgr.
add initialization of file manager related classes to FileMgrC.__init__.
add initialization of the two playlist opener classes to
FileMgrC.__init__.

book_reader.glade:
Remove the playlist open/create widgets from the file. They are now
managed in their own template file.

book_reader_view.py:
Remove classes NewBookOpenerV, ExistingBookOpenerM, and
ExistingBookOpenerV. They are now implemented in file_mgr_view.py

file_mgr_view.py:
add the gtk,Template class FileManagerViewOuterT to the module
Add the classes PlaylistOpenerView, ExistingPlaylistOpenerM, and
NavigationView.
Do some minor cleanup.

pinned_books_view.py:
transition to signal_.GLOBAL_TRANSMITTER.

pinned_books.py:
PinnedBooksC no longer acts as a signal relay for PinnedBooksV.
PinnedBooksC no longer inherits signal_.Signal.

signal_.py:
Add GLOBAL_TRANSMITTER instance to the module.

book_mark.py:
create file and add BookMark and related code from book_ease.py

file_mgr.ui:
Create file.
Contains template widgets for the file manager components.
