+++ src/gui/gtk/pinned_books_view.py
@@ -95,9 +95,6 @@ class PinnedBooksVC:
         self.pinned_books_view.remove_button.connect('clicked', self.on_control_button_clicked)
 
         self.pinned_books_vm.load_pinned_list(self.pinned_books_model.get_pinned_playlists())
-        # callback system
-        self.transmitter = signal_.Signal()
-        self.transmitter.add_signal('open_book')
 
     def get_pinned_list_new(self) -> Gtk.ListStore:
         """
@@ -117,7 +114,7 @@ class PinnedBooksVC:
         model, paths = sel.get_selected_rows()  # pylint: disable=unused-variable
         if paths:
             playlist_data = self.pinned_books_vm.get_playlist_data(paths[0])
-            self.transmitter.send('open_book', playlist_data)
+            signal_.GLOBAL_TRANSMITTER.send('open_book', playlist_data)
 
     def remove_selected_playlist(self):
         """remove the first playlist selected in the view from the pinned playlists"""
